+++ test/OtherOperators/OtherOperators.test.ts
@@ -1,0 +1,5 @@
+import RangeOperatorsTests from "./RangeOperatorsTests";
+
+describe("Other operators", () => {
+  describe("Range operators", RangeOperatorsTests.test);
+});