+++ test/StandardFunctions/OtherFunctionsTests.ts
@@ -1,0 +1,19 @@
+import TestConfiguration from "../TestConfiguration";
+import TestHelper from "../TestHelper";
+
+
+
+export default class OtherFunctionsTests {
+  static test() {
+    const tests: TestConfiguration[] = [
+      // log
+      new TestConfiguration("log(4)", undefined),
+
+      // print
+      new TestConfiguration("print(4)", 4),
+      new TestConfiguration('print("¡Hola, mundo!")', "¡Hola, mundo!"),
+    ];
+
+    TestHelper.runTests(tests);
+  }
+}