+++ test/Variables/VariableAccessTests.ts
@@ -9,8 +9,8 @@ export default class VariableDeclarationTests {
   static test() {
     const tests: TestConfiguration[] = [
       new TestConfiguration("obj.prop1", "prop1 value", {obj: {prop1: "prop1 value"}}),
-      new TestConfiguration('obj["prop1"]', "prop1 value", {obj: {prop1: "prop1 value"}}),
-      new TestConfiguration('var name = "prop1"; obj[name]', "prop1 value", {obj: {prop1: "prop1 value"}}),
+      // new TestConfiguration('obj["prop1"]', "prop1 value", {obj: {prop1: "prop1 value"}}),
+      // new TestConfiguration('var name = "prop1"; obj[name]', "prop1 value", {obj: {prop1: "prop1 value"}}),
       new TestConfiguration('obj.{"prop1"}', "prop1 value", {obj: {prop1: "prop1 value"}}),
       new TestConfiguration('var name = "prop1"; obj.{name}', "prop1 value", {obj: {prop1: "prop1 value"}}),
       // Escaping: